Refactor resourcestore package
* Remove dependency on `git.Checkout` and rename `checkoutManager` to
  `fileResourceStore`
* Inline configfilemanager and manifestfilemanager into the new
  `fileResourceStore`
